feat(#200): 겜구 매칭/ 맞춤 매칭 단계까지 비회원 접근 허용
